Optimized code separation and organization using blueprints and helpers function
Create blueprints folder. The api is separated into 4 different files, one for getting activities,
one for creating activities, one for creating subarrays such as competence array, and one for
user functionalities. In addition the helpers file is used for various mappings between the data coming from the
frontend and the data that are saved into the database.